Added support for other 1D hist types.

+++ rootmacro2rad.h
@@ -3,4 +3,4 @@
 #include <string.h>
 #define S32K   32768
 
-float hist[S32K]; //internal histogram
+double hist[S32K]; //internal histogram
